Update PlayfulParticles
Added a Round function to fix floating point errors. Fixed the conditionals in the Move functions because the if/else with greater than/less than was giving biases. Changed the particle from a circle to a line drawn between the particle position and its next position to give a better visual of the vector. Fixed a bug where the current particle wasn't being included in its own flock math.